pyproject.toml
[tool.poetry]
name = "django-auth"
version = "0.1.0"
description = ""
authors = ["SARAH AHMAD HUDAIB <SARAHHUDAIB@YAHOO.COM>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
